Cards display appropriately in teh Perk manager

+++ server/src/main/java/com/example/repository/PerkRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 import com.example.model.Perk;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PerkRepository extends CrudRepository<Perk, Integer> {
+public interface PerkRepository extends CrudRepository<Perk, Long> {
 }
